+++ chronosphere/tfschema/log_search_filter.go
@@ -1,0 +1,10 @@
+package tfschema
+
+import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+var LogSearchFilterSchema = map[string]*schema.Schema{
+	"query": {
+		Type:     schema.TypeString,
+		Required: true,
+	},
+}